feat(SDK): fallback now reverts to internal unity support
Previously, the Fallback SDK would just return null information and
not actually function. In this new feature, the internal Unity
VR support is used in the fallback so basic VR functionality can
still be provided if a headset that doesn't have SDK support can
still be used if it's compatible with `UnityEngine.VR`.